# docs/reference-guides/openapi.yaml
openapi: 3.1.0
info:
  title: packageR HTTP API
  version: next
  description: |
    Supported packageR endpoints for object operations, profile settings,
    configured shares, public packages, presigned URLs, and STAC catalogs.

    A deployment can add a path prefix with `PACKAGE_R_BASEURL`. Object path
    segments must be URL-encoded while `/` remains the path separator.
servers:
  - url: http://127.0.0.1:8888
    description: Local packageR server
...
paths:
...
  /api/resources/{path}:
    parameters:
      - $ref: "#/components/parameters/ObjectPath"
    get:
      tags: [Resources]
      operationId: getResource
      summary: List a directory or get file metadata
      security:
        - PackageRToken: []
      parameters:
        - $ref: "#/components/parameters/Checksum"
        - $ref: "#/components/parameters/Presign"
        - $ref: "#/components/parameters/FollowRedirect"
      responses:
        "200":
          description: File metadata or a directory listing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "307":
          $ref: "#/components/responses/PresignedRedirect"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Resources]
      operationId: createResource
      summary: Upload a file or create a directory
      description: A trailing `/` in the object path creates a directory.
      security:
        - PackageRToken: []
      parameters:
        - $ref: "#/components/parameters/Override"
      requestBody:
        required: false
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Resource created.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Resources]
      operationId: replaceResource
      summary: Replace an existing file
      security:
        - PackageRToken: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: File replaced.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Resources]
      operationId: copyOrMoveResource
      summary: Copy or move an object
      security:
        - PackageRToken: []
      parameters:
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum: [copy, rename]
        - name: destination
          in: query
          required: true
          description: URL-encoded destination path.
          schema:
            type: string
        - $ref: "#/components/parameters/Override"
      responses:
        "200":
          description: Copy or move completed.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Resources]
      operationId: deleteResource
      summary: Delete a file or directory tree
      security:
        - PackageRToken: []
      responses:
        "204":
          description: Resource deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    PackageRToken:
      type: apiKey
      in: header
      name: X-Auth
      description: Signed packageR session token returned by `/api/login`.
    SharePassword:
      type: apiKey
      in: header
      name: X-SHARE-PASSWORD
      description: URL-encoded password for a protected public share.
  parameters:
    ObjectPath:
      name: path
      in: path
      required: true
      description: Object path below the authenticated user's scope.
      schema:
        type: string
      example: catalog-sample/catalog.parquet
...
    Checksum:
      name: checksum
      in: query
      required: false
      schema:
        type: string
        enum: [md5, sha1, sha256, sha512]
    Presign:
      name: presign
      in: query
      required: false
      schema:
        type: boolean
        default: false
    FollowRedirect:
      name: followRedirect
      in: query
      required: false
      description: Return a 307 redirect when `presign` is true.
      schema:
        type: boolean
        default: false
...
    Override:
      name: override
      in: query
      required: false
      description: Replace an existing destination when permitted.
      schema:
        type: boolean
        default: false
...
  responses:
...
    PresignedRedirect:
      description: Temporary redirect to object storage.
      headers:
        Location:
          description: Presigned object-storage URL.
          schema:
            type: string
            format: uri
    BadRequest:
      description: Invalid request.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated identity cannot perform the action.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    MethodNotAllowed:
      description: Method is not valid for this resource.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Destination already exists.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Sorting:
      type: object
      additionalProperties: false
      required: [by, asc]
      properties:
        by:
          type: string
          enum: [name, size, modified]
        asc:
          type: boolean
...
    Resource:
      type: object
      required:
        - path
        - name
        - size
        - extension
        - modified
        - mode
        - isDir
        - isSymlink
        - type
      properties:
        path:
          type: string
        name:
          type: string
        size:
          type: integer
          format: int64
        extension:
          type: string
        modified:
          type: string
          format: date-time
        mode:
          type: integer
        isDir:
          type: boolean
        isSymlink:
          type: boolean
        type:
          type: string
        content:
          type: string
        checksums:
          type: object
          additionalProperties:
            type: string
        resolution:
          $ref: "#/components/schemas/ImageResolution"
        presignedURL:
          type: string
          format: uri
        stacBrowserURL:
          type: string
          format: uri
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        numDirs:
          type: integer
        numFiles:
          type: integer
        sorting:
          $ref: "#/components/schemas/Sorting"
    ImageResolution:
      type: object
      additionalProperties: false
      required: [width, height]
      properties:
        width:
          type: integer
        height:
          type: integer
...
    Error:
      type: string
      example: 404 Not Found